+++ lib/grid/armadillo_helpers.hpp
@@ -24,7 +24,7 @@
 #include <dune/common/fvector.hh>
 
 // Internal implementations for general Armadillo objects
-template<typename M, typename T, size_t size>
+template<typename M, typename T, int size>
 bool _armadillo_fieldvector_compare(const M& x, const Dune::FieldVector<T, size>& y)
 {
     if (x.n_rows != size)
@@ -35,7 +35,7 @@ bool _armadillo_fieldvector_compare(const M& x, const Dune::FieldVector<T, size>
     return true;
 }
 
-template<typename M, typename T, unsigned int rows, unsigned int cols>
+template<typename M, typename T, int rows, int cols>
 bool _armadillo_fieldmatrix_compare(const M& x, const Dune::FieldMatrix<T, rows, cols>& y)
 {
     if (x.n_rows != rows || x.n_cols != cols)
@@ -47,37 +47,37 @@ bool _armadillo_fieldmatrix_compare(const M& x, const Dune::FieldMatrix<T, rows,
     return true;
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const arma::Col<T>& x, const Dune::FieldVector<T, size>& y)
 {
     return _armadillo_fieldvector_compare(x, y);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const Dune::FieldVector<T, size>& x, const arma::Col<T>& y)
 {
     return _armadillo_fieldvector_compare(y, x);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const arma::subview_col<T>& x, const Dune::FieldVector<T, size>& y)
 {
     return _armadillo_fieldvector_compare(x, y);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const Dune::FieldVector<T, size>& x, const arma::subview_col<T>& y)
 {
     return _armadillo_fieldvector_compare(y, x);
 }
 
-template <typename T, size_t rows, size_t cols>
+template <typename T, int rows, int cols>
 bool operator==(const arma::Mat<T>& x, const Dune::FieldMatrix<T, rows, cols>& y)
 {
     return _armadillo_fieldmatrix_compare(x, y);
 }
 
-template <typename T, size_t rows, size_t cols>
+template <typename T, int rows, int cols>
 bool operator==(const Dune::FieldMatrix<T, rows, cols>& x, const arma::Mat<T>& y)
 {
     return _armadillo_fieldmatrix_compare(y, x);
@@ -94,37 +94,37 @@ namespace test_tools
 namespace tt_detail
 {
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const ::arma::Col<T>& x, const ::Dune::FieldVector<T, size>& y)
 {
     return ::operator==(x, y);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const ::Dune::FieldVector<T, size>& x, const ::arma::Col<T>& y)
 {
     return ::operator==(x, y);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const arma::subview_col<T>& x, const Dune::FieldVector<T, size>& y)
 {
     return ::operator==(x, y);
 }
 
-template <typename T, size_t size>
+template <typename T, int size>
 bool operator==(const Dune::FieldVector<T, size>& x, const arma::subview_col<T>& y)
 {
     return ::operator==(x, y);
 }
 
-template <typename T, size_t rows, size_t cols>
+template <typename T, int rows, int cols>
 bool operator==(const arma::Mat<T>& x, const Dune::FieldMatrix<T, rows, cols>& y)
 {
     return ::operator==(x, y);
 }
 
-template <typename T, size_t rows, size_t cols>
+template <typename T, int rows, int cols>
 bool operator==(const Dune::FieldMatrix<T, rows, cols>& x, const arma::Mat<T>& y)
 {
     return ::operator==(x, y);
